[tests] Remove tests that used blender to execute Albam.
The intention was good: run the blender executable and compare the
original file with the one exported. This was adding too much complexity
for little gain.
The tests that deal with research, e.g. understanding the formats,
should be separated from the implementations. Ideally, Blender testing
is not needed if a standart import/export intermediate format is used
(e.g. glTF), which is something I want to aim for.