Initial explicit import via file-explorer
Before: importing an arc file would search for all mod files inside of it,
uncompress and import them.
Now: arc files are initially loaded, an their internal structure can be
seen in a tree-like fashion, similar to the Blender outliner. Then one
can import each mod file individually.
This opens the possibility to import other types of files, and provide
arguments to them. For example, animations that need to be applied to an
armature.