Fix random fails to import/save in mtfw
Leave comment as a reminder no to change a certain variable
Looks like Using EnumProperty in Blender is buggy without storing the string.
There is a warning in the docks of a known bug in EnumProperty, but only for
callbacks[1]. In here, using file_item.app_id was unreliable; at certain
point while loading new files to the file_explorer the app_id changed to
the default one (re0).

[1] https://docs.blender.org/api/3.6/bpy.props.html#bpy.props.EnumProperty