Initial error handling improvement
Errors are now handled for the following operators:

* albam.import
* albam.export
* albam.patch
* albam.pack

When an exception happens, instead of showing the default long
traceback, a custom popup explaining what happens appears.
The traceback is only printed in the console, along with some extra
information, ready for copy-pasting into an issue or chat.
The popup includes links to the albam discord channel and github issues
as well.

More operators and further improvements should come in future
iterations.