Add export checks for material and textures (#87)
* Initial implementation of checks before export

When an exception of type `AlbamCheckFailure` is raised in operators
covered by albam_error_handler, extra details and solutions are
provided. The header changes as well, since the error is not unexpected.

* Initial export check for mtfw shader group

A check is performed before export that all materials use the mtfw shader group
on mtfw apps. If not, the affected materials are shown with instructions to fix it:

* Initial export check for dds textures

A check is performed before export that all textures present in the
materials assigned to the meshes to export are DDS only.
The wrong textures are reported in a popup with the materials where they
are used