Initial copy/paste operators for props (#93)
* Align custom props and buttons
Now they follow Blender's UI style

* Add export albam custom properties button
Store custom properties for the current app in a json file.
This json has  main key which is the name of the custom props used by
the app, to be used when importing.

* Add import custom properties button
Loads a json file and if it's valid and contains a key named like the
current custom props (i.e. the custom props for the app), load their
properties.
Some validation is done of the file, but not that individual properties
actually exist in the custom props yet.