Vfs refactor (#99)
Refactor de-duplicating a lot of code, deleting things that didn't make sense and renaming around.
Perhaps the most important refactor was that virtual files can be obtained by just a path so a lot of conversion and knowledge of the vfs is not needed in callers. Before an ugly knowledge of how paths were stored internally was necessary.
This is important because now multiple arcs can be combined very easily, given that in many cases the references are distributed among more than one arc file. The need came from RE6 where many mrl files are stored in different arcs from the main characters. But it will be useful, for example, for RE5 stages where times textures were missing for the same reason.

As a bonus, icons were added to the file list, so this also closes #81