Fix wrong bone positions in some imported models
Models that include rotation data in their inverse bind matrices were
being wrongly imported. If rotation is present, we don't do anything with it,
so probably that's the reason why some mirrored bones are not perfectly aligned.
Not a big issue.
So far only `rev2::uPl2100Head.arc::data/.../pl2103.mod` was observed to
have some sort of small rotation.