RE6 Import support (#103)
* Initial import support for RE6. Export enabled but tested only on one mesh

Export roundtrip shows minor weight problems in pl0000, but hangs the
game with just the mesh so far

* Use hash as default material name if available

This will allow better testing of generated mrls, since hashes will be
the same when exported. If hashes are not used (e.g. re5) we just keep
the traditional prefixes with the material index

* Fix issue with some weights importing

Some bones that were not supposed to have weights were imported due to
rounding errors. Discovered via tests