Add a tool - Vertex groups merger (#113)
This inconvenience was a thing from Reloaded times. Different costumes for the same character could have some variety, for example, Jill Battlesuit has 2 breast bones, while Jill BSAA only 1, so to port the same mesh for different costumes you need to merge two vertex groups into one.
Unfortunately the Blender way to do it with modifiers is cumbersome. 
It was tolerable for RE5 but porting RE5 characters to other games became a challenge because of how over-complicated the RE5 skeleton is in comparison with later games.

This new tool quickly merges 2 vertex groups into one.