Attempt to workaround segfault
In bpy==4.1 there's a segfault whenever any PropertyGroup class has a
method defined in it. Before a message was printed about unfreed memory
blocks.
Even if tests fail, the exitcode 1 will be overriden by 139, so for now
we are losing automatic failure detection of tests.
This is a first attempt to at least make both tests in 3.6 and 4.1 pass.
It can probably we improved further without needing to refactor all
PropertyGroup sub-classes yet.