Initial Blender 4.0+ support (#114)
Initial support for Blender 4.0 and above, keeping 3.6 compatibility as well.

Tests will run for bpy version 3.6 and 4.1

Note: In bpy==4.1 there's a segfault whenever any PropertyGroup class has a
method defined in it. Before a message was printed about un-freed memory
blocks.
Even if tests fail, the exitcode 1 will be overridden by 139, so for now
we are losing automatic failure detection of tests.